SQL을 JPA로 사용하기 및 데이터 수정 코드

+++ src/main/java/com/JtoP/Spring/boundedContext/answer/repository/AnswerRepository.java
@@ -3,5 +3,5 @@ package com.JtoP.Spring.boundedContext.answer.repository;
 import com.JtoP.Spring.boundedContext.answer.entity.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnswerRepository extends JpaRepository<Answer, Long> {
+public interface AnswerRepository extends JpaRepository<Answer, Integer> {
 }
